+++ Packages/UGF.Assemblies/Runtime/AssemblyBrowsableAttribute.cs
@@ -5,7 +5,7 @@ namespace UGF.Assemblies.Runtime
     /// <summary>
     /// Attribute used to mark specific assembly as available for types search.
     /// </summary>
-    [AttributeUsage(AttributeTargets.Assembly)]
+    [AttributeUsage(AttributeTargets.Assembly, AllowMultiple = true)]
     public sealed class AssemblyBrowsableAttribute : Attribute
     {
     }