+++ changelog.md
@@ -26,6 +26,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing.
 
+## 1.0.1 - 2019-03-32
+- [Commits](https://github.com/unity-game-framework/ugf-assemblies/compare/1.0.1...1.1.0)
+- [Milestone](https://github.com/unity-game-framework/ugf-assemblies/milestone/3?closed=1)
+
+### Added
+- `AssemblyEditorUtility` to control assembly attributes.
+- Context menu to `Assembly Definition File` to toggle usage of the `AssemblyBrowsableAttribute`.
+
+### Changed
+- `AssemblyBrowsableAttribute`: `AttributeUsage.AllowMultiple` set to `true` to allow multiple usage of the attribute on assembly in project.
+
 ## 1.0.1 - 2019-03-14
 - [Commits](https://github.com/unity-game-framework/ugf-assemblies/compare/1.0.0...1.0.1)
 - [Milestone](https://github.com/unity-game-framework/ugf-assemblies/milestone/2?closed=1)